guide and general cleanup
