+++ cli/src/modules/mod.rs
@@ -30,7 +30,7 @@ pub mod rpc;
 pub mod select;
 pub mod send;
 pub mod server;
-pub mod set;
+pub mod settings;
 pub mod sign;
 pub mod start;
 pub mod stop;
@@ -52,8 +52,9 @@ pub fn register_handlers(cli: &Arc<KaspaCli>) -> Result<()> {
         cli.handlers(),
         [
             account, address, close, connect, details, disconnect, estimate, exit, export, guide, help, history, import, rpc, list,
-            miner, monitor, mute, network, node, open, ping, reload, select, send, server, set, start, stop, sweep, theme, track,
-            transfer, wallet,
+            miner, monitor, mute, network, node, open, ping, reload, select, send, server, settings, sweep, track, transfer,
+            wallet,
+            // theme,  start, stop
         ]
     );
